Fixed orphaned java processes
This mainly fixes a problem where the script restarts but leaves
orphaned java jetty processes behind.  SQLForce registers an atexit()
handler to shut them down, but because we re-exec the main process,
those handlers are not called and their registration is lost in the
newly exec'd script.

The solution I've set up is to call the SQLForceServer.killServer()
method directly (which causes the java process to exit), then do
some child process reaping so that we don't leave zombies behind.

Also changed the "set" command so that "set var=val" works since I
always type that.

Added a "disconnect" command that logs out of the current profile.
Not very useful, but there for completeness.

Modified the SELECT command to only print the data table if rows
were returned.