Allow multiline commands
* use cmd2 instead of cmd
* minor changes to colors
* empty line doesn't do anything (would previously repeat last command)
* introduced table2sql helper function

The table2sql function allows pasting of a list of strings and will
combine them into one or more IN("a','b','c') style SQL constructs.
This is useful if you have a list of record identifiers and want to
perform a query (or update) using them.

table2sql breaks the list into groups of 250 to avoid generating SQL
statements that are too long.